requirements.txt
requests==2.31.0
pytest==8.1.1
pytest-allure-adaptor==1.7.0
allure-pytest==2.13.3